:ambulance: 잠 잘 잤다 진짜 다 죽었다 진짜

+++ src/main/java/com/pard/pree_be/feedback/report/dto/ReportRequestDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ReportRequestDto {
     private String name; // Metric name
-    private double counter; // Raw metric value
+    private int counter; // Raw metric value
     private int score; // Calculated score
     private String feedbackMessage; // Feedback
 }
